moving network debugger into its own class.
